🎉 version(v1.0.1): 更新版本
1. 实现充能技能, 兼容测试服焚影圣诀技改.

2. 开放 customLua API.

3. 提升内置循环对延迟的兼容性, 现在可使用内置循环计算 20~90 延迟. 同步更新示例 lua.

4. 开放 effect 武器·橙武.

5. 后端计算置信区间.

+++ source/include/frame/character/property/damage.h
@@ -1,11 +1,12 @@
 #ifndef FRAME_CHARACTER_PROPERTY_DAMAGE_H_
 #define FRAME_CHARACTER_PROPERTY_DAMAGE_H_
 
+#include <cstdint>
 #include <vector>
 
 namespace ns_frame {
 
-using event_tick_t = unsigned long long;
+using event_tick_t = int64_t;
 
 enum class DamageType {
     Physics,
