+++ source/include/frame/event.h
@@ -1,12 +1,13 @@
 #ifndef FRAME_EVENT_H_
 #define FRAME_EVENT_H_
 
+#include <cstdint>
 #include <set>
 #include <tuple>
 
 namespace ns_frame {
 
-using event_tick_t = unsigned long long;
+using event_tick_t = int64_t;
 using event_func_t = void (*)(void *, void *);
 
 /**
